Added audit functions to the database (#161)
* Added audit functions to the database

No table is audited yet: in order to audit e.g. the table `users` a
superuser can run:

    select audit.start('users');

This creates a table called `audit."public.users"` which contains all
the changes to the table, together with a couple of audit fields
(`audit_ts`, `audit_action`).

Different audit fields can be configured: see
[the docs](https://github.com/dvarrazzo/pgaudit/blob/master/doc/pgaudit.md)
for details.

* Added audit users and individuals tables

Added an audit field to record the `application_name` in the table. It
is a bit of a hack: the `audit.user_id` was designed for it. However
this would require configuring `postgres.conf`, which is not possible on
RDS.